# src/main/java/com/smartexpiry/item/api/CreateItemRequest.java
package com.smartexpiry.item.api;

import com.smartexpiry.expiry.domain.ShelfLifeUnit;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Positive;
import jakarta.validation.constraints.PositiveOrZero;
import jakarta.validation.constraints.Size;

import java.math.BigDecimal;
import java.time.LocalDate;

public record CreateItemRequest(
    @jakarta.validation.constraints.Pattern(regexp = "[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}") String clientId,
    @NotBlank @Size(max = 120) String name,
    @NotBlank String categoryId,
    @Size(max = 120) String brand,
    @PositiveOrZero BigDecimal quantity,
    @Size(max = 32) String unit,
    LocalDate productionDate,
    LocalDate expiryDate,
    @Positive Integer shelfLifeValue,
    ShelfLifeUnit shelfLifeUnit,
    LocalDate openedDate,
    @Positive Integer afterOpenValue,
    ShelfLifeUnit afterOpenUnit
) {}